requirements.txt
openai
pdfplumber
yfinance
python-dotenv
python-dateutil